Pdfpeaks/scripts/requirements.txt
pdf2docx
pdfplumber
openpyxl
pymupdf
python-pptx
pikepdf
Pillow
